feat: Quiz생성 플로우에 Knn search와 GPT 개발 유사도 파이프라인을 추가한다

+++ src/main/kotlin/org/gitanimals/quiz/infra/similarity/Tokenizer.kt
@@ -5,7 +5,7 @@ import org.springframework.web.service.annotation.PostExchange
 fun interface Tokenizer {
 
     @PostExchange("/pipeline/feature-extraction/sentence-transformers/all-mpnet-base-v2")
-    fun tokenize(request: Request): List<Double>
+    fun tokenize(request: Request): List<Float>
 
     data class Request(
         val inputs: String,
